requirements.txt
Django==2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
PyMySQL==0.10.0